Refactor of endpoints Moved PUT endpoints to user controller In the future the PUT admin endpoint will contain the whole AppUser object change
