+++ src/main/java/com/github/WeeiaEduTeam/InfinityFinanceAPI/appuser/AppUserUserController.java
@@ -1,14 +1,21 @@
 package com.github.WeeiaEduTeam.InfinityFinanceAPI.appuser;
 
-import com.github.WeeiaEduTeam.InfinityFinanceAPI.appuser.dto.AppUserDTO;
-import com.github.WeeiaEduTeam.InfinityFinanceAPI.appuser.dto.CreateAppUserAdminDTO;
-import com.github.WeeiaEduTeam.InfinityFinanceAPI.appuser.dto.CreateAppUserUserDTO;
+import com.github.WeeiaEduTeam.InfinityFinanceAPI.appuser.dto.*;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.net.URI;
+
+/*
+ * If you hit the
+ * - DELETE "/users" endpoint
+ * - PUT "/users/credentials" endpoint
+ *  make sure that you logout
+ *  current user to ensure getting
+ *  troubleshoot errors.
+ */
 
 @RestController
 @RequiredArgsConstructor
@@ -33,12 +40,6 @@ public class AppUserUserController {
         return ResponseEntity.created(URI.create("/users/" + user.getId())).body(user);
     }
 
-    /*
-    * If you hit the endpoint make sure
-    * that you logout current user to
-    * ensure getting troubleshoot errors.
-    */
-
     @DeleteMapping("/users")
     ResponseEntity<Void> deleteCurrentLoggedUser() {
         appUserUserService.deleteCurrentLoggedUser();
@@ -46,6 +47,22 @@ public class AppUserUserController {
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
     }
 
+    @PutMapping("/users/credentials")
+    ResponseEntity<AppUserDTO> replaceUserCredentials(
+            @RequestBody AppUserCredentialsDTO appUserCredentialsDTO) {
 
+        var replacedUser = appUserUserService.replaceUserCredentials(appUserCredentialsDTO);
+
+        return ResponseEntity.ok(replacedUser);
+    }
+
+    @PutMapping("/users/details")
+    ResponseEntity<AppUserDTO> replaceUserDetails(
+            @RequestBody ReplaceAppUserByUserDTO replaceAppUserByUserDTO) {
+
+        var replacedUser = appUserUserService.replaceUserDetails(replaceAppUserByUserDTO);
+
+        return ResponseEntity.ok(replacedUser);
+    }
 
 }
